Upgraded to python 3. Changed backend. Changed config.

+++ daemon.py
@@ -1,7 +1,8 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 import sys, os, time, atexit
 from signal import SIGTERM
+from datetime import datetime
 
 class Daemon:
     def __init__(self, pidfile, stdin='/dev/null', stdout='/dev/null', stderr='/dev/null'):
@@ -16,7 +17,7 @@ class Daemon:
             if pid > 0:
                 # exit first parent
                 sys.exit(0)
-        except OSError, e:
+        except OSError as e:
             sys.stderr.write("fork #1 failed: %d (%s)\n" % (e.errno, e.strerror))
             sys.exit(1)
 
@@ -31,27 +32,28 @@ class Daemon:
             if pid > 0:
                 # exit from second parent
                 sys.exit(0)
-        except OSError, e:
+        except OSError as e:
             sys.stderr.write("fork #2 failed: %d (%s)\n" % (e.errno, e.strerror))
             sys.exit(1)
 
         # redirect standard file descriptors
         sys.stdout.flush()
         sys.stderr.flush()
-        si = file(self.stdin, 'r')
-        so = file(self.stdout, 'a+')
-        se = file(self.stderr, 'a+', 0)
-        os.dup2(si.fileno(), sys.stdin.fileno())
+        si = open(self.stdin, 'r')
+        so = open(self.stdout, 'a+')
+        se = open(self.stderr, 'a+')
+        os.dup2(si.fileno(), sys.stdin.fileno())            
         os.dup2(so.fileno(), sys.stdout.fileno())
         os.dup2(se.fileno(), sys.stderr.fileno())
         
         # write pidfile
         atexit.register(self.delpid)
         pid = str(os.getpid())
-        file(self.pidfile,'w+').write("%s\n" % pid)
+        open(self.pidfile,'w+').write("%s\n" % pid)
 
     def delpid(self):
-        os.remove(self.pidfile)
+        if os.path.exist(self.pidfile):
+            os.remove(self.pidfile)
 
     def start(self):
         """
@@ -59,9 +61,8 @@ class Daemon:
         """
         # Check for a pidfile to see if the daemon already runs
         try:
-            pf = file(self.pidfile,'r')
-            pid = int(pf.read().strip())
-            pf.close()
+            with open(self.pidfile,'r') as pf:
+                pid = int(pf.read().strip())
         except IOError:
             pid = None
 
@@ -71,6 +72,8 @@ class Daemon:
             sys.exit(1)
 
         # Start the daemon
+        dt = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+        print('Start at ', dt)
         self.daemonize()
         self.run()
 
@@ -78,11 +81,12 @@ class Daemon:
         """
         Stop the daemon
         """
+        dt = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+        print('Stop at ', dt)
         # Get the pid from the pidfile
         try:
-            pf = file(self.pidfile,'r')
-            pid = int(pf.read().strip())
-            pf.close()
+            with open(self.pidfile,'r') as pf:
+                pid = int(pf.read().strip())
         except IOError:
             pid = None
 
@@ -96,22 +100,26 @@ class Daemon:
             while 1:
                 os.kill(pid, SIGTERM)
                 time.sleep(0.1)
-        except OSError, err:
+        except OSError as err:
             err = str(err)
             if err.find("No such process") > 0:
                 if os.path.exists(self.pidfile):
                     os.remove(self.pidfile)
             else:
-                print str(err)
+                print(str(err))
                 sys.exit(1)
 
     def restart(self):
         """
         Restart the daemon
         """
+        dt = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
         self.stop()
         self.start()
+        print('Restarted ', dt)
 
     def run(self):
+        """
+        To be overloaded
         """
